move `tools/legacy_keyring.py` to `chia/legacy/keyring.py` (#17252)
* move `tools/legacy_keyring.py` to `chia/legacy/keyring.py`

* init
